Rename lcm->jobmanager; make everything work
