+++ Preview/src/kubelayer_test.go
@@ -4,37 +4,18 @@ import (
 	"errors"
 	"testing"
 
-	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/kubernetes/fake"
 )
 
-func TestGetConfig(t *testing.T) {
-	GetConfig()
-}
-
 func TestApi(t *testing.T) {
-	var config, err = GetConfig()
-	if err != nil {
-		t.Errorf("unable to create config")
-	}
-	PingApi(config)
-
-	_, err = kubernetes.NewForConfig(config)
-	if err != nil {
-		t.Errorf("Unable to create clientset")
-	}
-
+	clientset := fake.NewSimpleClientset()
+	PingApi(clientset)
 }
 
 func TestCreateJobKubeLayer(t *testing.T) {
-	var config, _ = GetConfig()
-	var clientset, err = kubernetes.NewForConfig(config)
-	if err != nil {
-		t.Errorf("unable to create clientset")
-	}
-	_, err = CreateJob(clientset)
+	clientset := fake.NewSimpleClientset()
+	_, err := CreateJob(clientset)
 	if err != nil {
 		panic(errors.New("Create job failed."))
 	}
 }
-
-//TODO this test should use mocks
